Update allJavadocs task to auto-generate Javadoc links
Included a new dependency in the allJavadocs task of the build.gradle.kts file. This enables automatic generation of Javadoc links for all projects, improving documentation quality and completeness. A change was also made to use 'allJavadocs' task for the Github Actions workflow, to produce combined Javadocs for all projects, thus streamlining maintenance and review.